refactor(mon-pix): move QCM proposals
