+++ mon-pix/app/components/proposals/qcm-proposals.hbs
