+++ mon-pix/tests/integration/components/proposals/qcm-proposals_test.js
@@ -2,13 +2,13 @@ import { render } from '@1024pix/ember-testing-library';
 import { hbs } from 'ember-cli-htmlbars';
 import { module, test } from 'qunit';
 
-import setupIntlRenderingTest from '../../helpers/setup-intl-rendering';
+import setupIntlRenderingTest from '../../../helpers/setup-intl-rendering';
 
 module('Integration | Component | QCM proposals', function (hooks) {
   setupIntlRenderingTest(hooks);
 
   test('renders', async function (assert) {
-    await render(hbs`<QcmProposals />`);
+    await render(hbs`<Proposals::QcmProposals />`);
     assert.dom('.qcm-proposals').exists();
   });
 });
